+++ website/frontend/src/components/pages/Mailbox/Mailbox.tsx
@@ -4,7 +4,6 @@ import { Hex } from '@gear-js/api';
 import { Box } from 'layout/Box/Box';
 import { MailItem } from './children/MailItem/MailItem';
 import { Mail } from './types';
-import { getMails } from './helpers';
 import { LOCAL_STORAGE } from 'consts';
 import styles from './Mailbox.module.scss';
 
@@ -12,10 +11,27 @@ export const Mailbox: FC = () => {
   const [api] = useApi();
   const [mails, setMails] = useState<Mail[]>([]);
 
-  useEffect(() => {
+  const getMails = async () => {
     const publicKey = localStorage.getItem(LOCAL_STORAGE.PUBLIC_KEY_RAW) as Hex;
 
-    getMails(api, publicKey, setMails);
+    try {
+      await api.mailbox.subscribe(publicKey, (data: any) => {
+        const mailsHuman = data.toHuman();
+        const mailsArray: Mail[] = [];
+
+        for (const key in mailsHuman) {
+          mailsArray.push(mailsHuman[key]);
+        }
+
+        setMails(mailsArray);
+      });
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  useEffect(() => {
+    getMails();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
